backend/pyproject.toml
[project]
name = "ukcrm"
version = "0.1.0"
description = "ukcrm"
requires-python = ">=3.14"
dependencies = [
    "django>=5.2.8",
    "django-cors-headers>=4.9.0",
    "django-ordered-model>=3.7.4",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.29.0",
    "psycopg2-binary>=2.9.11",
    "python-decouple>=3.8",
]
